refactor(app): extract provider construction into internal/app (PR1a)
First step of the unified-bootstrap work (dev-docs/unified-session-bootstrap.md):
give provider-client construction and the provider→agent.Sender adapter a
single home so every entry point shares it instead of duplicating it.

- New internal/app: NewSender(SenderOptions) builds the vendor client and wraps
  it as an agent.Sender; DefaultBaseURL exposes the vendor endpoints for display.
  The providerSender adapter + replyFromResponse move here verbatim.
- cmd/octo: buildProvider splits into resolveAPIKey (key resolution + the CLI
  setup-help text stays here, a CLI concern) and buildSender (delegates to
  app.NewSender). All four call sites — chat, channel, init, conduct — migrate.
- cmd/octo no longer imports internal/provider or the vendor packages at all
  (config.go's endpoint display now goes through app.DefaultBaseURL).
- Adapter tests move to internal/app/sender_test.go alongside the code.
- .octorules: internal/app is now the single provider-construction site;
  internal/server's duplicate is called out as pending migration (PR2).

Behavior-preserving — full suite green, go vet + gofmt clean.
